chore(compaction): #632 A2 step 2 — strip transform layer, wasm surface, test surgery
Delete the LayoutTransform mechanism from bus/candidate_runner.rs
(CompactTransform, FoldTransform, the LayoutTransform trait,
TransformOutcome, CandidatePlan::with_transform, and the transform
slot/plumbing in CandidatePlan/run_plan/run_candidate_field) while
keeping the produce -> validate -> measure -> verdict -> rank framework
itself, which the live RFC-068 celldb campaign
(crates/core/tests/celldb_template.rs) consumes unmodified.

Remove the compact_layout/fold_layout LayoutOptions fields and their
flag-gated branches from bus/layout.rs (build_bus_layout,
build_rotation_aware_row_layout[_selected]'s stale-flag guard,
FOLD_SEARCH_ENTITY_THRESHOLD, the PlacedMachineSignature test), plus
their only other struct-literal site in decomposition_search.rs's
inner-call LayoutOptions build (not called out by the original scope
map, found by `cargo build`) and layout.rs's now-orphaned
cover_unpowered helper (was only called from the deleted fold-seam
pole-coverage repair inside compaction.rs; zero remaining callers).

Remove the same two Option<String> params (and their wiring into
LayoutOptions) from the three #[wasm_bindgen] entry points (layout,
layout_traced, layout_streaming) and the shared layout_options()
builder in wasm-bindings/src/lib.rs — a breaking change to the wasm
JS API that the web/ commit (next) follows up.

Test-file surgery:
- Delete crates/core/tests/fold_layout_knob.rs (whole file).
- crates/core/tests/candidate_runner.rs: delete the two fold/compact
  parity tests (one was the suite's slowest test at >60s) and port the
  two synthetic-transform ranking/verdict tests
  (BboxDoublingTransform/ShrinkOntoOverlapTransform) to synthetic
  DecompositionCandidate base producers instead, since what they
  exercise ("a candidate whose score/verdict differs from the
  incumbent's") never depended on the transform abstraction.
- crates/core/tests/objective.rs: delete the fold-ranks-above-native
  calibration test; keep measurement_is_total_or_it_refuses, ported off
  compact_layout onto a plain build.
- crates/core/tests/connectivity_parity.rs: delete the five
  compaction-transform-dependent tests (phase2_prefilter_*,
  phase2b_fold_prefilter_measurement, compaction_remaps_effective_rows)
  — all pinned compact_validated_geometry's/search_snake_fold's OWN
  internal admission-filter correctness, not K65-1/K65-2 in general.
  K65-1/K65-2 coverage on every non-compaction path is untouched.

crates/core/tests/cell_composition.rs (the RFC-057 fold/manifold-
routing test block) is NOT part of this commit — it needs its own
surgery pass, next.

Part of #632 A2 (owner call — session goal "complete B5 and A2 as
described").
